Start writing support for all block states.
This will parse the generated registry and generate an internal representation of the
block state that can be manipulated with get/set methods. In future, it will be able to
parse a block state as a String to operate in reverse.

This is intended as a bridge between the protocol IDs and the stable palette strings,
as well as a representation that allows me to change state fields (such as take state X,
change property A to B, then get the protocol ID of that new state, motivated by blocks
that connect. This should allow me to add connections to nearby states and get the new
protocol IDs to use for those new states.

This is pretty early days, but I can pretty print all the states already.