Prevent ConcurrentModification condition in ChunkDispatcher
ChunkDispatcher, by virtue of calling sendChunk, was clashing with
Player's loadCorrectChunks reading of the same variable. This commit
swaps the Set<> for a ConcurrentHashMap<> (Though that may not be the
final solution) to resolve this condition.

I also resolve a pathological case where if you are moving faster than
ChunkDispatcher could send chunk packets to the client, the same chunk
would be requested multiple times since loadCorrectChunks would still
see it as unloaded and request it be loaded again.