EntityPosition: Major improvements to position storage
This changes EntityPosition to just store a gosh darn double for pitch
and yaw instead of laundering back and forth all the time.

It also makes the internal storage format be Radians instead of Degrees
- you ask for Degrees explicitly if you want them.

To accomplish this I've moved the APIs in EntityPosition over to
properties with getters and have left deprecated methods in their place.

While cleaning up references in Javaland this incidentally cleans up
UpdateViewPosition / ChunkPosition packet handling a bunch, and I wish
that could be a second commit but I'm not unfixing this just to make a
bad version post-API-change just to do that.